AI Manager/requirements.txt
azure-core==1.23.1
azure-identity==1.9.0
azure-storage-file-share==12.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
dnspython==2.2.1
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sklearn == 0.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
